GLSL simplifications, and desaturation prevention.
I made the GLSL functions use 'inout' instead of returning a value, since I seem
to always be doing something like 'color = function(color, other thing)'. May as
well have 'color' be 'inout' instead of returning a separate value.

I also rescaled the output values to prevent unwanted desaturation, by making
them darker if the color transformation would cause any of the RGB primaries to
go out of gamut, but only out of gamut because of how bright they otherwise
would become - not because of their chromaticity coordinates.

I do this because my current gamut scaling function will desaturate colors that
are too bright, even if the chromaticity coordinates for that color are in-gamut
according to the triangle.