Switched to 'max normalization'
This makes it easier to see where the white point is, though I should probably
be drawing a dot as part of the 'drawPrimaries' function instead.

I'd have to modify it a bit more to really do what I want, though:

1. Pass a colorspace parameter in, which would contain the primaries.
2. Pass a color in to determine what color the triangle and dot should be.
3. Use the primaries held within the colorspace instead of primaries directly.
4. Determine how to draw the dot.
   a. Should it have a hole in the middle to see the color?
   b. Should it be fairly large, or just a single pixel?
   c. Best math for centering right on the whitepoint.

I'd also like the triangle (and theoretical dot) for each colorspace to have a
different color from the other one. Needless to say, that's all just on the TODO
list for another time.